add a "reflected XSS" vulnerability
This commit adds a URL parameter, 'name', that is used in a
welcome message if it exists. The name is unsanitized and so
can include JS and thus cause an XSS that runs on the client
side but does not exist on the server side (i.e., a "reflected
XSS").

The behavior of the system is unchanged if you go to the bare URL,
e.g., http://127.0.0.1:5000/ -- however, if you add a 'name' parameter
to the URL, e.g., http://127.0.0.1:5000?name=Peter it will display
a welcome message between "XSS Demo" and "Read, Search, and Post
Comments".

The name parameter can contain JS, e.g.:
http://127.0.0.1:5000?name=Peter<script>alert("boo")</script>

... and this JS will run in the page.